Move details dialog to a separate file
