+++ components/result-table/warnings.jsx
@@ -1,5 +1,18 @@
+import { useState } from "react";
+import { makeStyles } from "@material-ui/core/styles";
+
 import { Warning as WarningIcon } from "@material-ui/icons";
 import { Popover, Typography } from "@material-ui/core";
+
+
+const useStyles = makeStyles((theme) => ({
+  popover: {
+    pointerEvents: "none",
+  },
+  paper: {
+    padding: theme.spacing(1),
+  },
+}));
 
 export function WarningsPopover(props) {
   const [anchorEl, setAnchorEl] = useState(null);
